fix: surface nested migration errors in add-missing-system-fields command
When workspaceMigrationRunnerService.run() throws a
WorkspaceMigrationRunnerException, the error message only says
"Migration action 'create' for 'fieldMetadata' failed" without the
actual cause. The command now enriches the message with the nested
error details (metadata, workspaceSchema, actionTranspilation) before
re-throwing, matching how other commands surface errors.

Made-with: Cursor
